Repackage API into separate folder
Similar to how it is done with AE
