+++ build.gradle
@@ -51,6 +51,14 @@ jar {
 	})
 }
 
+sourceSets {
+	main {
+		java {
+			srcDirs += 'src/api/java'
+		}
+	}
+}
+
 processResources {
 	// this will ensure that this task is redone when the versions change.
 	inputs.property "mod_version", project.mod_version
