+++ gradle.properties
@@ -5,4 +5,13 @@ ccl_version=1.1.3.127
 ccc_version=1.0.4.29
 multipart_version=1.1.2.330
 nei_version=1.0.3.+
-mod_version=0.2-pr0
+mod_version=0.2-pr1
+
+#########################################################
+# API Stubs                                             #
+# Note: Do not edit these version numbers as they are   #
+# not the version of the mod, but stub code for AE2 to  #
+# compile                                               #
+#########################################################
+api_ic2_version=1
+api_rf_version=2
