+++ src/api/java/org/squiddev/cctweaks/api/IWorldPosition.java
