+++ src/api/java/org/squiddev/cctweaks/api/network/INetworkNode.java
