+++ src/api/java/org/squiddev/cctweaks/api/network/NetworkVisitor.java
