+++ src/api/java/org/squiddev/cctweaks/api/package-info.java
@@ -1,0 +1,4 @@
+@API(apiVersion = "${mod_version}", owner = "CCTweaks", provides = "CCTweaks|API")
+package org.squiddev.cctweaks.api;
+
+import cpw.mods.fml.common.API;
