+++ src/api/java/org/squiddev/cctweaks/api/turtle/ITurtleFuelProvider.java
