+++ src/api/java/org/squiddev/cctweaks/api/turtle/TurtleFuelRegistry.java
