+++ src/main/java/org/squiddev/cctweaks/core/registry/RefuelRegisters.java
@@ -41,35 +41,5 @@ public class RefuelRegisters implements IRegisterable{
 				return fuelToGive;
 			}
 		});
-
-		// IEnergyContainerItem
-		TurtleFuelRegistry.addFuelProvider(new ITurtleFuelProvider() {
-			@Override
-			public boolean canRefuel(ITurtleAccess turtle, ItemStack stack, int limit) {
-				return Config.config.turtleFluxRefuelEnable && stack != null && stack.getItem() instanceof IEnergyContainerItem;
-			}
-
-			@Override
-			public int refuel(ITurtleAccess turtle, ItemStack stack, int limit) {
-				int fluxAmount = Config.config.turtleFluxRefuelAmount;
-
-				// Avoid over refueling
-				int maxRefuel = turtle.getFuelLimit() - turtle.getFuelLevel();
-				int fluxLimit = (limit >= 64 ? maxRefuel : limit) * fluxAmount;
-
-				IEnergyContainerItem container = (IEnergyContainerItem) stack.getItem();
-
-				// Don't want to pull to much
-				fluxLimit = Math.min(container.getMaxEnergyStored(stack), fluxLimit);
-
-				int change = 1;
-				int progress = 0;
-				while (progress < fluxLimit && change > 0) {
-					change = container.extractEnergy(stack, fluxLimit - progress, false);
-					progress += change;
-				}
-				return progress / fluxAmount;
-			}
-		});
 	}
 }
